backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.4
pydantic[email]==2.7.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
python-multipart==0.0.9

# AI
google-generativeai==0.7.2

# HTTP client
httpx==0.27.0

# Data processing
pandas==2.2.2
openpyxl==3.1.4
xlrd==2.0.1

# External DB drivers (install as needed)
aiomysql==0.2.0
# snowflake-connector-python==3.10.0  # heavy — install separately
# firebase-admin==6.5.0               # install separately

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
